ai-service/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.35.0
python-multipart==0.0.22
Pillow==12.2.0
pydantic==2.10.6
opencv-python-headless==4.11.0.86
numpy==2.2.6
